Renamed messages from "option" to "method" to make things clearer Added better support for tab completion and ignoring the first arg of falling blocks and item disguises, so /disguise fallingblock setburning - Is now possible Fixed itemstack parsing returning illegal/unwanted air block

+++ src/main/java/me/libraryaddict/disguise/disguisetypes/MiscDisguise.java
@@ -131,7 +131,7 @@ public class MiscDisguise extends TargetedDisguise {
     public int getData() {
         switch (getType()) {
             case FALLING_BLOCK:
-                return (int) ((FallingBlockWatcher) getWatcher()).getBlock().getDurability();
+                return ((FallingBlockWatcher) getWatcher()).getBlock().getDurability();
             case PAINTING:
                 return ((PaintingWatcher) getWatcher()).getArt().getId();
             case SPLASH_POTION:
